flask_mysql_project_mysqlDB/app/requirements.txt
flask
mysql-connector-python
requests
